feat(backend): overhaul entity, dto, and service layers for enhanced features
- Removed `PropertyPhotoRepository` as no longer needed.
- Introduced DTOs for property exporting, profile updates, and saved properties (`SavedPropertyExportDto`, `UserExportDto`, etc.).
- Added `PropertySearchCriteria` for cleaner search handling.
- Enhanced translation handling by replacing legacy fields with a centralized structure and validation (`applyTranslations`).
- Updated validation constraints (e.g., size limits) for DTOs.
- Optimized media (photos, plans) handling with dedicated helper methods.
- Logging added for key lifecycle actions (e.g., property creation, user profile updates).

+++ backend/src/main/java/com/app/backend/dto/PropertyFilter.java
@@ -13,12 +13,12 @@ public record PropertyFilter(
         List<Integer> rooms,
         BigDecimal m2Min,
         BigDecimal m2Max,
-        Integer floorMin,
-        Integer floorMax,
+        Short floorMin,
+        Short floorMax,
         Boolean notGround,
         Boolean notTop,
-        Integer yearMin,
-        Integer yearMax,
+        Short yearMin,
+        Short yearMax,
         List<String> features,
         String completion
 ) {}
